backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
python-dateutil==2.8.2
httpx==0.26.0
anthropic>=0.25.0
PyJWT[crypto]>=2.8.0
